requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
python-dotenv>=1.0.0
openrouter
langchain-openrouter